+++ plugins/10_Dragonflight/localization/frFR.lua
@@ -920,6 +920,7 @@ L['smelly_treasure_chest_label'] = 'Coffre au trésor kifouette'
 L['options_icons_smelly_treasure_chest'] = 'Coffres au trésor kifouette'
 L['options_icons_smelly_treasure_chest_desc'] = 'Afficher les emplacements possibles des {object:Coffres au trésor kifouette}.'
 
+L['seething_cache'] = 'Cache bouillonnante'
 L['seething_cache_treasure_note'] = 'Pour pouvoir voir la {object:Cache bouillonnante} et récupérer le {item:192779}, vous devrez obtenir 3x charges du debuff {spell:399342}, récupéré sur les {object:Orbes bouillonnants} situés à la {location:Caldeira Zaqali} dans la {location:Grotte de Zaralek}.'
 L['chest_of_the_flights_treasure_note'] = 'Pour ouvrir le coffre au trésor, vous devez cliquer sur les {object:Gemmes renforcées} dans l\'ordre {note:Rouge > Noire > Bleue > Jaune > Verte}.'
 L['curious_top_hat_note'] = 'Interagissez avec le {npc:205010} pendant que vous avez le buff {spell:410288} pour obtenir {item:205021}. Si vous n\'avez pas le buff et que vous vous approchez de lui, il s\'enfuira.'