+++ plugins/10_Dragonflight/zones/zaralek_cavern.lua
@@ -490,9 +490,9 @@ map.nodes[43058256] = Treasure({
 
 map.nodes[32333935] = Treasure({
     quest = 73410,
+    label = L['seething_cache'],
     note = L['seething_cache_treasure_note'],
     rewards = {
-        Achievement({id = 17786, criteria = 59220}), -- Treasures of Zaralek Cavern
         Mount({item = 192779, id = 1623}) -- Seething Slug
     },
     pois = {